[x86] perf/x86: Use event_base_rdpmc for the RDPMC userspace support
Message-id: <20210108001544.23913-2-mpetlan@redhat.com>
Patchwork-id: 347310
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 001/197] perf/x86: Use event_base_rdpmc for the RDPMC userspace support
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit 75608cb02ea5dd997990e2998eca3670cb71a18c
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Thu Jul 23 10:11:04 2020 -0700

description
===========
The RDPMC index is always re-calculated for the RDPMC userspace support,
which is unnecessary.

The RDPMC index value is stored in the variable event_base_rdpmc for
the kernel usage, which can be used for RDPMC userspace support as well.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>