[x86] perf/x86/intel: Fix the name of perf METRICS
Message-id: <20210108001544.23913-6-mpetlan@redhat.com>
Patchwork-id: 347314
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 005/197] perf/x86/intel: Fix the name of perf METRICS
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit bbdbde2a415d9f479803266cae6fb0c1a9f6c80e
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Thu Jul 23 10:11:08 2020 -0700

description
===========
Bit 15 of the PERF_CAPABILITIES MSR indicates that the perf METRICS
feature is supported. The perf METRICS is not a PEBS feature.

Rename pebs_metrics_available perf_metrics.

The bit is not used in the current code. It will be used in a later
patch.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>