[x86] perf/x86/intel: Use switch in intel_pmu_disable/enable_event
Message-id: <20210108001544.23913-7-mpetlan@redhat.com>
Patchwork-id: 347315
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 006/197] perf/x86/intel: Use switch in intel_pmu_disable/enable_event
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit 58da7dbe6f036fefe504a4bb452afbd39bba73f7
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Thu Jul 23 10:11:09 2020 -0700

description
===========
Currently, the if-else is used in the intel_pmu_disable/enable_event to
check the type of an event. It works well, but with more and more types
added later, e.g., perf metrics, compared to the switch statement, the
if-else may impair the readability of the code.

There is no harm to use the switch statement to replace the if-else
here. Also, some optimizing compilers may compile a switch statement
into a jump-table which is more efficient than if-else for a large
number of cases. The performance gain may not be observed for now,
because the number of cases is only 5, but the benefits may be observed
with more and more types added in the future.

Use switch to replace the if-else in the intel_pmu_disable/enable_event.

If the idx is invalid, print a warning.

For the case INTEL_PMC_IDX_FIXED_BTS in intel_pmu_disable_event, don't
need to check the event->attr.precise_ip. Use return for the case.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>