[kernel] perf/core: Fix a memory leak in perf_event_parse_addr_filter()
Message-id: <20210108001544.23913-45-mpetlan@redhat.com>
Patchwork-id: 347353
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 044/197] perf/core: Fix a memory leak in perf_event_parse_addr_filter()
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit 7bdb157cdebbf95a1cd94ed2e01b338714075d00
Author: kiyin(尹亮) <kiyin@tencent.com>
Date: Wed Nov 4 08:23:22 2020 +0300

description
===========
As shown through runtime testing, the "filename" allocation is not
always freed in perf_event_parse_addr_filter().

There are three possible ways that this could happen:

 - It could be allocated twice on subsequent iterations through the loop,
 - or leaked on the success path,
 - or on the failure path.

Clean up the code flow to make it obvious that 'filename' is always
freed in the reallocation path and in the two return paths as well.

We rely on the fact that kfree(NULL) is NOP and filename is initialized
with NULL.

This fixes the leak. No other side effects expected.

[ Dan Carpenter: cleaned up the code flow & added a changelog. ]
[ Ingo Molnar: updated the changelog some more. ]

Fixes: 375637bc5249 ("perf/core: Introduce address range filtering")
--
 kernel/events/core.c | 12 +++++-------
 1 file changed, 5 insertions(+), 7 deletions(-)

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>