[tools] perf tools: Make GTK2 support opt-in
Message-id: <20210108001544.23913-80-mpetlan@redhat.com>
Patchwork-id: 347388
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 079/197] perf tools: Make GTK2 support opt-in
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit 4751bddd3f983af2004ec470ca38b42d7a8a53bc
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Sep 4 17:11:59 2020 -0300

description
===========
This is bitrotting, nobody is stepping up to work on it, and since we
treat warnings as errors, feature detection is failing in its main,
faster test (tools/build/feature/test-all.c) because of the GTK+2
infobar check.

So make this opt-in, at some point ditch this if nobody volunteers to
take care of this.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>