[tools] perf parse-events: Reduce casts around bp_addr
Message-id: <20210108001544.23913-117-mpetlan@redhat.com>
Patchwork-id: 347425
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 118/197] perf parse-events: Reduce casts around bp_addr
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit aa98d8482c8397d4087248e1f50d50078b34326d
Author: Ian Rogers <irogers@google.com>
Date: Thu Sep 24 17:39:03 2020 -0700

description
===========
perf_event_attr bp_addr is a u64. parse-events.y parses it as a u64, but
casts it to a void* and then parse-events.c casts it back to a u64.
Rather than all the casts, change the type of the address to be a u64.

This removes an issue noted in:

  https://lore.kernel.org/lkml/20200903184359.GC3495158@kernel.org/

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>