[tools] perf streams: Compare two streams
Message-id: <20210108001544.23913-154-mpetlan@redhat.com>
Patchwork-id: 347463
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 155/197] perf streams: Compare two streams
Bugzilla: 1884818 1908305
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1908305
https://bugzilla.redhat.com/show_bug.cgi?id=1908305

upstream
========
commit 47ef8398c369fa77e0fdbadfebb3885879491721
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Oct 9 10:28:41 2020 +0800

description
===========
Stream is the branch history which is aggregated by the branch records
from perf samples. Now we support the callchain as stream.

If the callchain entries of one stream are fully matched with the
callchain entries of another stream, we think two streams are matched.

For example,

   cycles: 1, hits: 26.80%                 cycles: 1, hits: 27.30%
   -----------------------                 -----------------------
             main div.c:39                           main div.c:39
             main div.c:44                           main div.c:44

Above two streams are matched (we don't consider the case that source
code is changed).

The matching logic is, compare the chain string first. If it's not
matched, fallback to dso address comparison.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>