[fs] ext4: treat buffers with write errors as containing valid data
Message-id: <20210111145230.10575-15-lczerner@redhat.com>
Patchwork-id: 347692
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 fs PATCH 14/21] ext4: treat buffers with write errors as containing valid data
Bugzilla: 1908714
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: ZhangXiaoxu <zhangxiaoxu5@huawei.com>

BZ 1908714
https://bugzilla.redhat.com/show_bug.cgi?id=1908714

BREW 34183182
https://brewweb.devel.redhat.com/taskinfo?taskID=34183182

commit 7963e5ac901251c7a3b36fe7c987623a3f309393

I got some errors when I repair an ext4 volume which stacked by an
iscsi target:
    Entry 'test60' in / (2) has deleted/unused inode 73750.  Clear?
It can be reproduced when the network not good enough.

When I debug this I found ext4 will read entry buffer from disk and
the buffer is marked with write_io_error.

If the buffer is marked with write_io_error, it means it already
wroten to journal, and not checked out to disk. IOW, the journal
is newer than the data in disk.
If this journal record 'delete test60', it means the 'test60' still
on the disk metadata.

In this case, if we read the buffer from disk successfully and create
file continue, the new journal record will overwrite the journal
which record 'delete test60', then the entry corruptioned.

So, use the buffer rather than read from disk if the buffer is marked
with write_io_error.

Signed-off-by: Zhang Xiaoxu <zhangxiaoxu5@huawei.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>