[netdrv] ibmvnic: Harden device Command Response Queue handshake
Message-id: <20210114134814.13132-9-didoming@redhat.com>
Patchwork-id: 348833
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH BZ1901513 v2 0008/0047] ibmvnic: Harden device Command Response Queue handshake
Bugzilla: 1901513
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1901513
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=34245511

In some cases, the device or firmware may be busy when the
driver attempts to perform the CRQ initialization handshake.
If the partner is busy, the hypervisor will return the H_CLOSED
return code. The aim of this patch is that, if the device is not
ready, to query the device a number of times, with a small wait
time in between queries. If all initialization requests fail,
the driver will remain in a dormant state, awaiting a signal
from the device that it is ready for operation.

commit 36a782fdcc2bfe39265ac4640f29404493b52a55

Signed-off-by: Thomas Falcon <tlfalcon@linux.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Diego Domingos <didoming@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>