[arm64] kvm: arm64: Don't access PMCR_EL0 when no PMU is available
Message-id: <20210121142719.17604-10-eric.auger@redhat.com>
Patchwork-id: 350621
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 virt 1907826 890373 PATCH v2 09/21] KVM: arm64: Don't access PMCR_EL0 when no PMU is available
Bugzilla: 1890373 1907826
RH-Acked-by: Gavin Shan <gshan@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Marc Zyngier <maz@kernel.org>

We reset the guest's view of PMCR_EL0 unconditionally, based on
the host's view of this register. It is however legal for an
implementation not to provide any PMU, resulting in an UNDEF.

The obvious fix is to skip the reset of this shadow register
when no PMU is available, sidestepping the issue entirely.
If no PMU is available, the guest is not able to request
a virtual PMU anyway, so not doing nothing is the right thing
to do!

It is unlikely that this bug can hit any HW implementation
though, as they all provide a PMU. It has been found using nested
virt with the host KVM not implementing the PMU itself.

Fixes: ab9468340d2bc ("arm64: KVM: Add access handler for PMCR register")
Reviewed-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20201210083059.1277162-1-maz@kernel.org
(cherry picked from commit 2a5f1b67ec577fb1544b563086e0377f095f88e2)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>