[netdrv] net: stmmac: don't stop NAPI processing when dropping a packet
Message-id: <20210122175729.2891936-55-msalter@redhat.com>
Patchwork-id: 350988
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 054/332] net: stmmac: don't stop NAPI processing when dropping a packet
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit 07b3975352374c3f5ebb4a42ef0b253fe370542d
From: Aaro Koskinen <aaro.koskinen@nokia.com>

Currently, if we drop a packet, we exit from NAPI loop before the budget
is consumed. In some situations this will make the RX processing stall
e.g. when flood pinging the system with oversized packets, as the
errorneous packets are not dropped efficiently.

If we drop a packet, we should just continue to the next one as long as
the budget allows.

Signed-off-by: Aaro Koskinen <aaro.koskinen@nokia.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>