[netdrv] net: stmmac: Set Flow Control to automatic mode in the driver
Message-id: <20210122175729.2891936-62-msalter@redhat.com>
Patchwork-id: 351064
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 061/332] net: stmmac: Set Flow Control to automatic mode in the driver
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit e9989339063dd3a566b67caaa857e6077bfee2cb
From: Jose Abreu <Jose.Abreu@synopsys.com>

By default Flow Control feature is not being enabled in stmmac.

This is a useful feature that can prevent loss of packets and now that
XGMAC already supports it (along with GMAC and QoS) it makes sense to
activate it.

Switch the module parameter to FLOW_AUTO so that Flow Control is
activated.

Signed-off-by: Jose Abreu <joabreu@synopsys.com>
Cc: Joao Pinto <jpinto@synopsys.com>
Cc: David S. Miller <davem@davemloft.net>
Cc: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: Alexandre Torgue <alexandre.torgue@st.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>