[netdrv] net: stmmac: Fix possible deadlock when disabling EEE support
Message-id: <20210122175729.2891936-105-msalter@redhat.com>
Patchwork-id: 350961
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 104/332] net: stmmac: Fix possible deadlock when disabling EEE support
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit 0867bb9768dedaae2c0e572ade5ccaa682a9298c
From: Jon Hunter <jonathanh@nvidia.com>

When stmmac_eee_init() is called to disable EEE support, then the timer
for EEE support is stopped and we return from the function. Prior to
stopping the timer, a mutex was acquired but in this case it is never
released and so could cause a deadlock. Fix this by releasing the mutex
prior to returning from stmmax_eee_init() when stopping the EEE timer.

Fixes: 74371272f97f ("net: stmmac: Convert to phylink and remove phylib logic")
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Tested-by: Thierry Reding <treding@nvidia.com>
Acked-by: Willem de Bruijn <willemb@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>