[netdrv] net: stmmac: Fix page pool size
Message-id: <20210122175729.2891936-182-msalter@redhat.com>
Patchwork-id: 350993
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 181/332] net: stmmac: Fix page pool size
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit 4f28bd956e081fc018fe9b41ffa31573f17bfb61
From: Thierry Reding <treding@nvidia.com>

The size of individual pages in the page pool in given by an order. The
order is the binary logarithm of the number of pages that make up one of
the pages in the pool. However, the driver currently passes the number
of pages rather than the order, so it ends up wasting quite a bit of
memory.

Fix this by taking the binary logarithm and passing that in the order
field.

Fixes: 2af6106ae949 ("net: stmmac: Introducing support for Page Pool")
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>