[netdrv] net: stmmac: Do not stop PHY if WoL is enabled
Message-id: <20210122175729.2891936-187-msalter@redhat.com>
Patchwork-id: 350974
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 186/332] net: stmmac: Do not stop PHY if WoL is enabled
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit 3e2bf04fb0447aa4b967b8000125178f55ae7800
From: Jose Abreu <Jose.Abreu@synopsys.com>

If WoL is enabled we can't really stop the PHY, otherwise we will not
receive the WoL packet. Fix this by telling phylink that only the MAC is
down and only stop the PHY if WoL is not enabled.

Fixes: 74371272f97f ("net: stmmac: Convert to phylink and remove phylib logic")
Signed-off-by: Jose Abreu <Jose.Abreu@synopsys.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>