[netdrv] net: stmmac: xgmac: Fix RSS writing wrong keys
Message-id: <20210122175729.2891936-190-msalter@redhat.com>
Patchwork-id: 350978
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 189/332] net: stmmac: xgmac: Fix RSS writing wrong keys
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit 56627336b655494356a23e65200c21718a1c3b2b
From: Jose Abreu <Jose.Abreu@synopsys.com>

Commit b6b6cc9acd7b, changed the call to dwxgmac2_rss_write_reg()
passing it the variable cfg->key[i].

As key is an u8 but we write 32 bits at a time we need to cast it into
an u32 so that the correct key values are written. Notice that the for
loop already takes this into account so we don't try to write past the
keys size.

Fixes: b6b6cc9acd7b ("net: stmmac: selftest: avoid large stack usage")
Signed-off-by: Jose Abreu <Jose.Abreu@synopsys.com>
Reviewed-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>