[netdrv] net: stmmac: Rework TX Coalesce logic
Message-id: <20210122175729.2891936-227-msalter@redhat.com>
Patchwork-id: 351000
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 226/332] net: stmmac: Rework TX Coalesce logic
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit c2837423cb54551302f037106830ee607b4563e0
From: Jose Abreu <Jose.Abreu@synopsys.com>

Coalesce logic currently increments the number of packets and sets the
IC bit when the coalesced packets have passed a given limit. This does
not reflect very well what coalesce was meant for as we can have a large
number of packets that are coalesced and then a single one, sent later
on that has the IC bit.

Rework the logic so that it coalesces only upon a limit of packets and
sets the IC bit for large number of packets.

Signed-off-by: Jose Abreu <Jose.Abreu@synopsys.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>