[netdrv] net: stmmac: removed enabling eee in EEE set callback
Message-id: <20210122175729.2891936-304-msalter@redhat.com>
Patchwork-id: 351118
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 303/332] net: stmmac: removed enabling eee in EEE set callback
Bugzilla: 1870102
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1870102

commit 7241c5a697479c7d0c5a96595822cdab750d41ae
From: Voon Weifeng <weifeng.voon@intel.com>

EEE should be only be enabled during stmmac_mac_link_up() when the
link are up and being set up properly. set_eee should only do settings
configuration and disabling the eee.

Without this fix, turning on EEE using ethtool will return
"Operation not supported". This is due to the driver is in a dead loop
waiting for eee to be advertised in the for eee to be activated but the
driver will only configure the EEE advertisement after the eee is
activated.

Ethtool should only return "Operation not supported" if there is no EEE
capbility in the MAC controller.

Fixes: 8a7493e58ad6 ("net: stmmac: Fix a race in EEE enable callback")
Signed-off-by: Voon Weifeng <weifeng.voon@intel.com>
Acked-by: Mark Gross <mgross@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>