[redhat] fix genlog.py for patches with single line changelog
Message-id: <20210129211034.251787-1-herton@redhat.com>
Patchwork-id: 352134
Patchwork-instance: patchwork
O-Subject: [RHEL8 PATCH] [redhat] fix genlog.py for patches with single line changelog
Bugzilla:
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Jan Stancek <jstancek@redhat.com>

Bugzilla: INTERNAL
Upstream Status: RHEL only
Tested: verified output with single line changelog in a patch

The current version of genlog.py script looks for a patchwork id in a fixed
line, which is always there if patch was applied through patchwork. Also, for
patches we commit we will always have a changelog big enough to so it always
have more than 2 lines (we would not allow a patch without proper changelog).

However, anyone could locally apply a patch for testing with only eg. a single
line in the changelog (only the title, without the "body"), which breaks
genlog.py. Add an additional check to avoid it:

Traceback (most recent call last):
  File "(...)/redhat/genlog.py", line 97, in <module>
    log_item, cves, bugs, zbugs = parse_commit(c)
  File "(...)/redhat/genlog.py", line 58, in parse_commit
    patchwork = lines[2].startswith("Patchwork-id: ")

I also checked that git refuses commiting a patch with empty changelog,
so I think we are covered for this.

Reported-by: Luis Goncalves <lgoncalv@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>