[s390] s390/kdump: fix out-of-memory with PCI
Message-id: <20210211180719.519196-1-prudo@redhat.com>
Patchwork-id: 352939
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH] [s390] s390/kdump: fix out-of-memory with PCI
Bugzilla: 1927744
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Bugzilla: 1927744
Upstream Status: RHEL only
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=34893821
Tested: by IBM

This patch fixes a bug in the memory detection for the kdump kernel
which causes the memory_end to contain the total available memory of the
system instead of the available crashkernel memory. This can cause
problems as some subsystems scale their memory consumption on the
available memory given in memory_end, e.g. PCI for the IOMMU bitmaps.
This unnecessarily increases the memory usage in the kdump kernel
substantially and thus also the chance for it to run oom.

The fix restricts the memory for the kdump kernel as if the "mem="
parameter was passed on the kernel command line. If the "mem=" parameter
is passed on the command line its value takes precedence.

The bug was accidentally fixed upstream in 73045a08cf55 ("s390: unify
identity mapping limits handling") which is out of scope for a backport
into 8.4.

Signed-off-by: Philipp Rudo <prudo@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>