net/mlx5e: Disable performance optimization for IPv4/IPv6 ethertype
Bugzilla: http://bugzilla.redhat.com/1928671
Bugzilla: http://bugzilla.redhat.com/1919807
Upstream: RHEL-only

This effectively reverts the following commit with minimal diversion
from the upstream kernel:
f7cfe86dc950 ("[netdrv] net/mlx5e: Optimize performance for IPv4/IPv6 ethertype")

This patch introduced an optimization for IPv4/IPv6 ethertypes. However, it
also caused a regression where MPLS traffic can match a rule that looks
at 'ip' protocol only, which shouldn't match the MPLS traffic.

A proper upstream fix is still under discussions, but for now, in order to
fix the regression in RHEL, we are going to disable this optimization.

Reverting the upstream patch cannot be done cleanly, so instead we disable
the capability that the driver was looking for to make sure that it can
activate this optimization.

Signed-off-by: Alaa Hleihel <ahleihel@redhat.com>