vdpa/mlx5: Add VDPA driver for supported mlx5 devices
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1877270

Note: There're three features that are not backported to RHEL8, so
      below three were removed from the print_features():

      MLX5_LOG_VIO_FLAG(VIRTIO_NET_F_HASH_REPORT);
      MLX5_LOG_VIO_FLAG(VIRTIO_NET_F_RSS);
      MLX5_LOG_VIO_FLAG(VIRTIO_NET_F_RSC_EXT);

Add a front end VDPA driver that registers in the VDPA bus and provides
networking to a guest. The VDPA driver creates the necessary resources
on the VF it is driving such that data path will be offloaded.

Notifications are being communicated through the driver.

Currently, only VFs are supported. In subsequent patches we will have
devlink support to control which VF is used for VDPA and which function
is used for regular networking.

Reviewed-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Eli Cohen <eli@mellanox.com>
Link: https://lore.kernel.org/r/20200804162048.22587-13-eli@mellanox.com
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Acked-by: Jason Wang <jasowang@redhat.com>
(cherry picked from commit 1a86b377aa2147a7c866b03142e848c18e5f3cb8)
Signed-off-by: Jason Wang <jasowang@redhat.com>