redhat: use tags from git notes for zstream to generate changelog
Bugzilla: INTERNAL
Upstream Status: RHEL only
Tested: make rh-release for main and 8.3 branches and checked the
        generated changelog

v2: - Use maj.min* notes ref for zstream. Previously only maj.min ref was
      used. This change allows kernel-rt to use their notes along with the
      zstream ones for rt specific patches.

Previously for zstream we amended commit messages while applying patches from
patchwork and we added tags: Bugzilla: ybzs, Z-Bugzilla: zbzs and CVE: cves
if necessary. These were used by the genspec.sh to generate
spec changelog for zstreams.

Keeping the same approach with the gitlab workflow would mean to rebase
commits in merge requests for zstream, which is something we probably
don't want in the merge oriented workflow. Other approach would be to
require these tags in the merge request commits. This would be fine for
MRs created by maintainers, but could be confusing for others while
creating zstream specific MRs. This would also require changes in
bugzilla webhook.

To avoid any additional requirements or changes in genlog.py, let's just
use git notes to store zstream tags. The change is that zstream
maintainer will not amend commit messages, but will be adding the
required tags to git notes. With this change we will feed genlog.py with
git notes instead of commit messages for zstreams. There should be no
change for ystream.

Each zstream is expected to have its own ref for notes based on the major
and minor version. For example refs/notes/8.3 for 8.3.z. It's expected
that maintainer makes sure the notes are presented and up-to-date.
Fetching them automatically is IMHO not safe, because we cannot know the
context.

The format in git notes should be the same as expected by genlog.py.

 Bugzilla: ybzs
 Z-Bugzilla: zbzs
 CVE: cves

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>