time: Enable __kernel_timespec for 32-bit vdso build
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1936282
Upstream: RHEL only
Tested: With a test kernel built and booted up with this patch, running
	the command "podman run docker.io/i386/alpine:latest date"
	produced the correct result of "Wed Mar 10 18:56:01 UTC 2021".

Previous upstream time related backports enables the 32-bit vdso page
to add a new symbol "__vdso_clock_gettime64" which is supposed to
accept a __kernel_timespec parameter with 64-bit time. Unfortunately,
__kernel_timespec is currently aliased to timespec which has 32-bit
time when built in 32-bit mode. This causes a mismatch when an user
32-bit application happens to use the __vdso_clock_gettime64() vdso
function to retrieve a 64-bit __kernel_timespec structure but get a
32-bit timespec instead.

The upstream commit c70a772fda11 ("y2038: remove struct definition
redirects") which should be able to fix the timing code regression
reported in the BZ was not backported because it would break kABI.

Instead, the __kernel_timespec structure is now enabled when not in
64-bit mode. This affects only the building of the 32-bit vdso and
nothing else. This will fix the regression without causing kABI problem.

Signed-off-by: Waiman Long <longman@redhat.com>