scsi: target: iscsi: Redo iscsit_check_session_usage_count() return code
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1939458

The return value of iscsit_check_session_usage_count() is only checked if
it was not allowed to sleep. If it returns `2' then a timer is prepared. If
it returns something else or if it was allowed to sleep then it is ignored.

Let iscsit_check_session_usage_count() return true if it needs to arm the
timer - otherwise false. This simplifies the code flow of the only caller.

Link: https://lore.kernel.org/r/20201220203638.43615-4-bigeasy@linutronix.de
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit f88a10f80da9ed1ab1ba7496b70e9a0cdd8f7cf8)
Signed-off-by: Maurizio Lombardi <mlombard@redhat.com>