net: skbext: fix secpath use-after-free with ip fragmentation
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1901454
Upstream status: RHEL only
Tested: by the customer

This is a RHEL-only patch to the skb extensions system, since we
cannot convert secpath to use extensions as this would break kABI.

RHEL8's version of skb_ext_copy puts all "extensions" (extensions +
secpath) but doesn't reset the secpath pointer. This can lead to a
use-after-free if the skb that is copied to already had a secpath.

This patch modifies skb_ext_copy to also overwrite the secpath with
the source skb's.

We need to copy the implementation of secpath_get here, because
skbuff.h cannot include xfrm.h.

This patch is only needed in RHEL, since upstream has converted
secpath to skb extensions, so the secpath is already copied over.

Fixes: 81471773237a ("[include] sk_buff: fix skb_ext_put() and skb_ext_reset() calls")
Signed-off-by: Sabrina Dubroca <sdubroca@redhat.com>