cifs: revalidate mapping when we open files for SMB1 POSIX
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1933527
Tested: Manual testing and upstream buildbot

commit cee8f4f6fcabfdf229542926128e9874d19016d5

Under SMB1 + POSIX, if an inode is reused on a server after we have read and
cached a part of a file, when we then open the new file with the
re-cycled inode there is a chance that we may serve the old data out of cache
to the application.
This only happens for SMB1 (deprecated) and when posix are used.
The simplest solution to avoid this race is to force a revalidate
on smb1-posix open.

Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
    Reviewed-by: Paulo Alcantara (SUSE) <pc@cjr.nz>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit cee8f4f6fcabfdf229542926128e9874d19016d5)
    Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>