selinux: fix deadlock in security_set_bools()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1924230
Upstream Status: RHEL8-only

We must avoid calling audit_log() while holding the policy write lock,
because if there is an audit rule that filters based on SELinux context,
then audit_log() may end up calling selinux_audit_rule_match(), which
tries to take a policy read lock, causing a deadlock.

Upstream, this was fixed indirectly via a conversion of the rwlock to
RCU, but that change is too complex for this high-priority bugfix and
also introduced some new bugs, which are still being addressed upstream.

Therefore, fix this via a RHEL8-only patch that takes the audit_log()
calls outside the write critical section. This requires memorizing the
names of the changed booleans in an auxiliary array, so that the audit
messages can be constructed without touching the policydb structure when
outside the rwlock.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>