net: udp: make udp_tunnel_nic_info structure extensible
Bugzilla: https://bugzilla.redhat.com/1948450
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=36115938
Upstream Status: RHEL8 only

The structure udp_tunnel_nic_info was introduced recently in RHEL-8.4
and pointer to it is a part of net_device structure. This means it is
part of KABI and should be protected.

Because that pointer was added to net_device structure using RH_KABI_USE
macro the KABI checker is not able to detect unsafe modification of
udp_tunnel_nic_info struct. Any such modification will result in
silent KABI breaker.

The structure was not extended using RH_KABI_RESERVE so any additional
fields cannot be added using RH_KABI_USE. Additionally instances of
udp_tunnel_nic_info are usually created statically by NIC drivers so
usage of RH_KABI_EXTEND is also not possible.

To resolve this situation the structure is extended using
RH_KABI_AUX_EMBED and a special __RH_UDP_TUNNEL_NIC_INFO_EXTENDED
flag is introduced to indicate that the structure instance that
has this flag set in field udp_tunnel_nic_info.flags is an extended
one and it is safe to access its RHEL extension structure.

An example how a NIC driver should use this is embedded in a comment
in the header file.

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Petr Oros <poros@redhat.com>
Signed-off-by: Ivan Vecera <ivecera@redhat.com>