redhat/configs/editconfig: Add support for a bugzilla entry
Upstream Status: RHEL Only
Build Info: Not required - this is a script and is not part of the build
      	    process
Bugzilla: INTERNAL

All config changes should have bugzilla entries associated with them.
Allow the user to specify a bug number either via the command line (-b)
or via a menu entry.  Display a warning if no bugzilla number is provided.

Signed-off-by: David Arcari <darcari@redhat.com>

v2: always output a BZ line