SMB3: Fix mkdir when idsfromsid configured on mount
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1952263
Tested: Manual testing and upstream buildbot

commit c8c412f976124d85b8ded85c6ac3f760c12b63a3

mkdir uses a compounded create operation which was not setting
the security descriptor on create of a directory. Fix so
mkdir now sets the mode and owner info properly when idsfromsid
and modefromsid are configured on the mount.

    Signed-off-by: Steve French <stfrench@microsoft.com>
    CC: Stable <stable@vger.kernel.org> # v5.8
    Reviewed-by: Paulo Alcantara (SUSE) <pc@cjr.nz>
    Reviewed-by: Pavel Shilovsky <pshilov@microsoft.com>
    (cherry picked from commit c8c412f976124d85b8ded85c6ac3f760c12b63a3)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>