Handle STATUS_IO_TIMEOUT gracefully
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1952263
Tested: Manual testing and upstream buildbot

commit 8e670f77c4a55013db6d23b962f9bf6673a5e7b6

Currently STATUS_IO_TIMEOUT is not treated as retriable error.
It is currently mapped to ETIMEDOUT and returned to userspace
for most system calls. STATUS_IO_TIMEOUT is returned by server
in case of unavailability or throttling errors.

This patch will map the STATUS_IO_TIMEOUT to EAGAIN, so that it
can be retried. Also, added a check to drop the connection to
not overload the server in case of ongoing unavailability.

    Signed-off-by: Rohith Surabattula <rohiths@microsoft.com>
    Reviewed-by: Aurelien Aptel <aaptel@suse.com>
    Reviewed-by: Pavel Shilovsky <pshilov@microsoft.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 8e670f77c4a55013db6d23b962f9bf6673a5e7b6)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>