smb3: fix stat when special device file and mounted with modefromsid
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1952263
Tested: Manual testing and upstream buildbot

commit 3c3317daef0afa0cd541fc9c1bfd6ce8bbf1129a

When mounting with modefromsid mount option, it was possible to
get the error on stat of a fifo or char or block device:
        "cannot stat <filename>: Operation not supported"

Special devices can be stored as reparse points by some servers
(e.g. Windows NFS server and when using the SMB3.1.1 POSIX
Extensions) but when the modefromsid mount option is used
the client attempts to get the ACL for the file which requires
opening with OPEN_REPARSE_POINT create option.

    Signed-off-by: Steve French <stfrench@microsoft.com>
    CC: Stable <stable@vger.kernel.org>
    Reviewed-by: Ronnie Sahlberg <lsahlber@redhat.com>
    Reviewed-by: Shyam Prasad N <sprasad@microsoft.com>
    (cherry picked from commit 3c3317daef0afa0cd541fc9c1bfd6ce8bbf1129a)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>