smb3: do not try to cache root directory if dir leases not supported
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1952263
Tested: Manual testing and upstream buildbot

commit 3c6e65e679182d55779ef6f8582f0945af4319b0

To servers which do not support directory leases (e.g. Samba)
it is wasteful to try to open_shroot (ie attempt to cache the
root directory handle).  Skip attempt to open_shroot when
server does not indicate support for directory leases.

Cuts the number of requests on mount from 17 to 15, and
cuts the number of requests on stat of the root directory
from 4 to 3.

    Signed-off-by: Steve French <stfrench@microsoft.com>
    CC: Stable <stable@vger.kernel.org> # v5.1+
    (cherry picked from commit 3c6e65e679182d55779ef6f8582f0945af4319b0)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>