cifs: remove bogus debug code
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1952263
Tested: Manual testing and upstream buildbot

commit d367cb960ce88914898cbfa43645c2e43ede9465

The "end" pointer is either NULL or it points to the next byte to parse.
If there isn't a next byte then dereferencing "end" is an off-by-one out
of bounds error.  And, of course, if it's NULL that leads to an Oops.
Printing "*end" doesn't seem very useful so let's delete this code.

Also for the last debug statement, I noticed that it should be printing
"sequence_end" instead of "end" so fix that as well.

    Reported-by: Dominik Maier <dmaier@sect.tu-berlin.de>
    Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit d367cb960ce88914898cbfa43645c2e43ede9465)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>